Break up zone validator tests too
